Struct fts::walkdir::FileType

pub struct FileType {
    // some fields omitted
}

A file type of the directory entry like std::fs::FileType.

Methods

impl FileType

fn is_dir(&self) -> bool

Test whether this file type represents a directory.

fn is_file(&self) -> bool

Test whether this file type represents a regular file.

fn is_symlink(&self) -> bool

Test whether this file type represents a symbolic link.