Struct fts::walkdir::DirEntry

pub struct DirEntry {
    // some fields omitted
}

Methods

impl DirEntry

fn path(&self) -> &Path

Returns the full path to the file that this entry represents.

The full path is created by joining the original path to WalkDir::new with the filename of this entry.

fn metadata(&self) -> Option<Metadata>

Return the metadata for the file that this entry points at.

fn file_type(&self) -> FileType

Return the file type for the file that this entry points at.

fn file_name(&self) -> &OsStr

Returns the bare file name of this directory entry without any other leading path component.

fn depth(&self) -> usize

Returns the depth at which this entry was created relative to the original path to WalkDir::new.